Make NavTextLink an external component
Bring anchor tag inside NavIcon component

Move Nav comp into its own folder
