+++ pages/index.jsx
@@ -1,15 +1,7 @@
 import React from 'react';
 
-import Card from '../src/components/Card';
-
 const MainIndexPage = () => {
-  const gridWrapperStyle = {
-    display: 'flex',
-    justifyContent: 'center',
-  };
-
-  return (
-    <p>hello world</p>
-)};
+  return <p>this is temporary</p>;
+};
 
 export default MainIndexPage;
