+++ src/components/NavTextLink/styles.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const NavTextLinkContainer = styled.div`
+  display: inline-block;
+  margin-right: 2rem;
+
+  & > a {
+    text-decoration: none;
+  }
+`;
+
+export { NavTextLinkContainer };
